detect cycles in member relationships
Attempting to work with some of the AWS APIs (e.g. Sagemaker, DynamoDB)
was resulting in stack overflows. Turns out this was due to the
*shapeSpec.Schema() method, which is a recursive call, not detecting
cyclic relationships where member fields were of a type the same as the
parent. For instance, in DynamoDB's API, the AttributeValue shape is a
struct that has a member "M" (yes, that is the name of the field) which
is a map of string to AttributeValue. This cyclic relationship was
causing infinite recursion to happen in the Schema() function, as the
"M" field's Schema infinitely referred to the AttributeValue shape.

This patch fixes the infinite recursion issue by passing a slice of
string member shape names that are visited during Schema() method
inspection and upon finding a member shape in the visited list of shape
names, not attempting to re-determine the member's schema, instead just
planting a JSON reference in its place.